Refactor(NoteEditor): Use string resource for 'Paste Image'
This commit refactors the "Paste Image from Clipboard" text in the `NoteEditor` to use a string resource (`R.string.paste_image_from_clipboard`) instead of a hardcoded string.

This improves internationalization (i18n) by allowing the text to be translated.

Additionally, the `modifier` parameter in the `NoteEditorScreen` composable has been moved to be the first parameter, adhering to standard Jetpack Compose conventions.