feat: Set dynamic title for NoteEditor
This commit updates the `NoteEditor` to display a dynamic title in the toolbar, differentiating between adding a new note and editing an existing one.

Previously, the title was always "Edit note". Now, it will display:
- "Add note" when creating a new note.
- "Edit note" when modifying an existing one.

A new string resource, `cardeditor_title_add_note`, has been added to support this change.