feat(NoteEditor): Redesign TopAppBar primary actions
This commit redesigns the "Save" and "Preview" actions in the `NoteEditorTopBar`.

The "Save" action is now a filled `Button` with text, making it more prominent. The "Preview" action has been updated to use a new `visibility_24px` vector icon and is styled as a `FilledIconButton` to match the overflow menu button. The order of actions has also been reversed to "Overflow," "Preview," and "Save."