refactor(Decks): Add option to hide empty Default deck
This commit introduces a `skipEmptyDefault` parameter to `SelectableDeck.fromCollection` and `decks.allNamesAndIds`. This allows callers to optionally exclude the "Default" deck from the list of decks if it is empty.

The `CardBrowserViewModel` and `NoteEditorViewModel` have been updated to use this new parameter, simplifying the code by removing the manual filtering of the "Default" deck.