feat(NoteEditor): Pre-fill first field from intent text
This commit enhances the note-adding experience by automatically populating the first field of a new note when it's created via an `ACTION_PROCESS_TEXT` intent or a similar "share to AnkiDroid" action.

When the `NoteEditor` is launched to add a new note with incoming text, this text is now passed to the `NoteEditorViewModel`. The view model then sets this `initialFieldText` as the content of the first field in the newly created note, streamlining the process of adding content from other apps.

The logic in `NoteEditorFragment` has been updated to correctly extract this text from the intent and handle various note-adding entry points.