feat(CardBrowser): Improve keyboard handling for search expansion
This commit refactors the keyboard trigger mechanism within the `CardBrowserLayout`.

Previously, a `Boolean` state (`showKeyboard`) was used to request focus and show the software keyboard. This has been replaced with an integer `keyboardTrigger` that is incremented on each request.

This change prevents an extra recomposition that occurred when resetting the boolean state within the `LaunchedEffect`.